use exec to make chromium the first process in the container
Docker by default sends a signal to the process started by the
ENTRYPOINT statement (e.g. the first process in the container). In
this case this is the shell which executes start-browser.sh. The
shell will not forward some signals such as SIGINT (Ctrl+C). Also
docker stop seems not to work since it tries to use SIGTERM first
before using the force, from the journal while stopping:

  Container b44066006d5... failed to exit within 10 seconds of signal 15 - using the force"

Using exec makes chromium that process.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>