Introduce Cog as an alternative web browser engine
This commit adds Cog as an alternative to Chromium for a web browser
engine in the kiosk-mode-browser container.

Chromium is still the default browser engine used to open web pages,
but one can change to Cog by passing '--env KIOSK_LAUNCHER=cog' to
Docker when starting the container.

Currently, due to the lack of support on non-GPU devices (e.g. iMX7),
Cog is only installed on Wayland builds (iMX6 and iMX8).

Chromium and Cog tested on Apalis iMX6 and Verdin iMX8MM. Chromium
tested on Colibri iMX7.

Related-to: TOR-1890

Signed-off-by: Sergio Prado <sergio.prado@toradex.com>